+++ pkg/domain/blueprintSpec.go
@@ -73,20 +73,6 @@ type BlueprintConfiguration struct {
 	DryRun bool
 }
 
-type BlueprintUpgradePlan struct {
-	DogusToInstall   []string
-	DogusToUpgrade   []string
-	DogusToUninstall []string
-
-	ComponentsToInstall   []string
-	ComponentsToUpgrade   []string
-	ComponentsToUninstall []string
-
-	RegistryConfigToAdd    []string
-	RegistryConfigToUpdate []string
-	RegistryConfigToRemove []string
-}
-
 // ValidateStatically checks the blueprintSpec for semantic errors and sets the status to the result.
 // Here will be only checked, what can be checked without any external information, e.g. without dogu specification.
 // returns a domain.InvalidBlueprintError if blueprint is invalid
@@ -125,9 +111,12 @@ func (spec *BlueprintSpec) ValidateStatically() error {
 func (spec *BlueprintSpec) validateMaskAgainstBlueprint() error {
 	var errorList []error
 	for _, doguMask := range spec.BlueprintMask.Dogus {
-		dogu, noDoguFoundError := FindDoguByName(spec.Blueprint.Dogus, doguMask.Name.Name)
+		dogu, noDoguFoundError := FindDoguByName(spec.Blueprint.Dogus, doguMask.Name.SimpleName)
 		if noDoguFoundError != nil {
 			errorList = append(errorList, fmt.Errorf("dogu %q is missing in the blueprint", doguMask.Name))
+		}
+		if doguMask.TargetState == TargetStatePresent && dogu.TargetState == TargetStateAbsent {
+			errorList = append(errorList, fmt.Errorf("absent dogu %q cannot be present in blueprint mask", dogu.Name.SimpleName))
 		}
 		if !spec.Config.AllowDoguNamespaceSwitch && dogu.Name.Namespace != doguMask.Name.Namespace {
 			errorList = append(errorList, fmt.Errorf(
@@ -178,11 +167,15 @@ func (spec *BlueprintSpec) CalculateEffectiveBlueprint() error {
 	}
 
 	spec.EffectiveBlueprint = EffectiveBlueprint{
-		Dogus:                   effectiveDogus,
-		Components:              spec.Blueprint.Components,
-		RegistryConfig:          spec.Blueprint.RegistryConfig,
-		RegistryConfigAbsent:    spec.Blueprint.RegistryConfigAbsent,
-		RegistryConfigEncrypted: spec.Blueprint.RegistryConfigEncrypted,
+		Dogus:      effectiveDogus,
+		Components: spec.Blueprint.Components,
+		Config:     spec.Blueprint.Config,
+	}
+	validationError := spec.EffectiveBlueprint.validateOnlyConfigForDogusInBlueprint()
+	if validationError != nil {
+		spec.Status = StatusPhaseInvalid
+		spec.Events = append(spec.Events, BlueprintSpecInvalidEvent{ValidationError: validationError})
+		return validationError
 	}
 	spec.Status = StatusPhaseEffectiveBlueprintGenerated
 	spec.Events = append(spec.Events, EffectiveBlueprintCalculatedEvent{})
@@ -207,7 +200,7 @@ func (spec *BlueprintSpec) calculateEffectiveDogu(dogu Dogu) (Dogu, error) {
 		Version:     dogu.Version,
 		TargetState: dogu.TargetState,
 	}
-	maskDogu, noMaskDoguErr := spec.BlueprintMask.FindDoguByName(dogu.Name.Name)
+	maskDogu, noMaskDoguErr := spec.BlueprintMask.FindDoguByName(dogu.Name.SimpleName)
 	if noMaskDoguErr == nil {
 		emptyVersion := core.Version{}
 		if maskDogu.Version != emptyVersion {